Renamed dependencies
