+++ apps/web/src/features/product/ProductView.tsx
@@ -4,8 +4,8 @@ import styles from "./ProductView.module.css"
 import ProductImage from "@/assets/1.webp"
 import HeartIconStd from "@/assets/icons/heart-regular-full.svg?react"
 
-import ButtonStd from '../../ButtonStd'
-import Card from "../../Card"
+import ButtonStd from "@/components/ui/ButtonStd"
+import Card from "@/components/ui/Card"
 // import CommentCard from '@/components/CommentCard'
 
 const ProductView = () => {
