+++ apps/web/src/pages/ProductPage.tsx
@@ -1,11 +1,10 @@
 import React from 'react'
 import { useNavigate } from 'react-router-dom'
 
-import MainContainer from "../components/containers/MainContainer";
-import NavBar from "../components/ui/features/navbar/NavBar"
-import ProductCategories from "../components/ui/features/products/ProductCategories";
+import MainContainer from "@/components/containers/MainContainer";
+import ProductCategories from "@/features/products/ProductCategories";
 import ButtonStd from '@/components/ui/ButtonStd';
-import ProductView from "@/components/ui/features/product/ProductView"
+import ProductView from "@/features/product/ProductView"
 
 import BackIcon from "@/assets/icons/angle-left-solid-full.svg?react"
